connectly-api/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
confluent-kafka==2.6.2
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-db-connection-pool==1.2.5
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
hiredis==3.1.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
redis==5.2.1
requests==2.32.3
SQLAlchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
